style(bridge): rename directory types → interfaces
