+++ bridge/src/interfaces/contract-description.ts
